feat: add support for metadata field indexing
